Move the e2e tests in a separate package
This allows to avoid polluting the metallb go.mod with unnecessary
dependencies, and relegates the need for pinning (because of the e2e
framework) only to the e2e go.mod

On top of that, we clean the main go.mod file removing all the
dependencies that were necessary for tests only.
