+++ e2etest/pkg/metrics/prometheus.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 	"net/url"
 
-	"go.universe.tf/metallb/e2etest/pkg/executor"
+	"go.universe.tf/e2etest/pkg/executor"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clientset "k8s.io/client-go/kubernetes"
